refactor: move core domain logic to src/core
