+++ src/__tests__/audio.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 
-import { synth } from '../audio';
+import { synth } from '../core/audio';
 
 const createMockGainNode = () => ({
   connect: vi.fn(),
