+++ src/__tests__/polygons.test.ts
@@ -6,8 +6,8 @@ import {
   deduplicateAdjacentStrings,
   MAX_WRAP_OVERSHOOT,
 } from "../shapes/polygons";
-import { getFretboardNotes } from "../guitar";
-import { STANDARD_TUNING } from "../guitar";
+import { getFretboardNotes } from "../core/guitar";
+import { STANDARD_TUNING } from "../core/guitar";
 
 describe("buildPolygonFromNotes", () => {
   it("returns empty array when all strings have no notes (empty shape)", () => {
