+++ src/__tests__/practiceLens.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment jsdom
 import { describe, it, expect, beforeEach } from "vitest";
 import { createStore } from "jotai";
-import type { PracticeLens } from "../theory";
+import type { PracticeLens } from "../core/theory";
 import { k } from "../utils/storage";
 import {
   practiceLensAtom,
