+++ src/components/NoteGrid.tsx
@@ -1,6 +1,6 @@
 import { useRef, useEffect } from "react";
 import clsx from "clsx";
-import { formatAccidental, getNoteDisplay } from "../theory";
+import { formatAccidental, getNoteDisplay } from "../core/theory";
 import shared from "./shared.module.css";
 
 interface NoteGridProps {
