+++ src/core/audio.ts
