+++ src/core/degrees.ts
