+++ src/core/guitar.ts
