+++ src/core/theory.ts
