+++ src/core/theoryCatalog.ts
